web/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
numpy
pandas
finance-datareader
sqlmodel
SQLAlchemy
APScheduler
python-jose[cryptography]
passlib[bcrypt]
tenacity
yfinance

